+++ packages/layout-gpu/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @antv/layout-gpu
 
+## 1.1.5
+
+### Patch Changes
+
+- Use type module when creating WebWorker
+- Updated dependencies
+  - @antv/layout@1.2.4
+
+## 1.1.4
+
+### Patch Changes
+
+- Replace eventemitter with @antv/event-emitter
+- Updated dependencies
+  - @antv/layout@1.2.3
+
+## 1.1.3
+
+### Patch Changes
+
+- Add sideEffects field in package.json
+- Updated dependencies
+  - @antv/layout@1.2.2
+
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @antv/layout@1.2.1
+
 ## 1.1.1
 
 ### Patch Changes