+++ packages/layout/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @antv/layout
 
+## 1.2.4
+
+### Patch Changes
+
+- Use type module when creating WebWorker
+
+## 1.2.3
+
+### Patch Changes
+
+- Replace eventemitter with @antv/event-emitter
+
+## 1.2.2
+
+### Patch Changes
+
+- Add sideEffects field in package.json
+
+## 1.2.1
+
+### Patch Changes
+
+- Use comlink instead of webpack worker-loader
+
 ## 1.2.0
 
 ### Minor Changes